requirements.txt
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
python-binance>=1.0.19
